requirements.txt
scanpy==1.9.*
scvelo==0.2.*
anndata==0.8.0
numpy>=1.17.0
matplotlib>=3.4
pandas>=1.0
scipy>=1.4
scikit-misc==0.2.0
tslearn==0.5.3.*
scikit-learn
torchsde
seaborn
tqdm